save the orders for the client and display them for the admin

+++ flowershop/src/main/java/com/everis/flowershop/repository/OrderDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.everis.flowershop.repository.entity.Orders;
 
 @Repository
-public interface OrderDAO extends JpaRepository<Orders, Integer> {
+public interface OrderDAO extends JpaRepository<Orders, Long> {
 
 }
